GameFrame work
Continuing to work on the GameFrame class. So far, I've got the
bet/number of decks portions working properly, as well as adding an
image to the frame. Currently, I'm working on getting the last card
drawn to be added to the frame based on the location of the previous
card drawn.